+++ lib/src/connect.dart
@@ -25,7 +25,7 @@ abstract class HttpChannel {
    *
    * To start all channels, please use [StreamServer.stop] instead.
    */
-  void close();
+  Future close();
   /** Indicates whether the channel is closed.
    */
   bool get isClosed;